+++ include/unidade-de-area.h
@@ -1,0 +1,7 @@
+#ifndef UNIDADE_DE_AREA_H
+#define UNIDADE_DE_AREA_H
+
+void converterArea();
+
+#endif
+
