+++ include/unidade-de-massa.h
@@ -2,6 +2,5 @@
 #define UNIDADE_DE_MASSA_H
 
 void converterMassa();
-void converterArea();
 
 #endif
